src/1. Intro/pyproject.toml
[project]
name = "1-intro"
version = "0.1.1"
description = "Basic interactions with LLMs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
